safari-600.7-branch/LayoutTests/platform/gtk/fast/dom/Range/getClientRects-expected.txt

Test 1
PASS rects.length is 5
PASS rects[0].left is 8
PASS rects[0].top is 8
PASS rects[0].width is 400
PASS rects[0].height is 160
PASS rects[1].left is 8
PASS rects[1].top is 19
PASS rects[1].width is 396
FAIL rects[1].height should be 18. Was 17.
PASS rects[2].left is 8
PASS rects[2].top is 59
PASS rects[2].width is 398
FAIL rects[2].height should be 18. Was 17.
PASS rects[3].left is 8
PASS rects[3].top is 99
PASS rects[3].width is 360
FAIL rects[3].height should be 18. Was 17.
PASS rects[4].left is 8
PASS rects[4].top is 139
PASS rects[4].width is 306
FAIL rects[4].height should be 18. Was 17.
Test 2
PASS rects.length is 4
PASS rects[0].left is 8
FAIL rects[0].top should be 215. Was 213.
PASS rects[0].width is 396
FAIL rects[0].height should be 18. Was 17.
PASS rects[1].left is 8
FAIL rects[1].top should be 255. Was 253.
PASS rects[1].width is 398
FAIL rects[1].height should be 18. Was 17.
PASS rects[2].left is 8
FAIL rects[2].top should be 295. Was 293.
PASS rects[2].width is 360
FAIL rects[2].height should be 18. Was 17.
PASS rects[3].left is 8
FAIL rects[3].top should be 335. Was 333.
PASS rects[3].width is 306
FAIL rects[3].height should be 18. Was 17.
Test 3
PASS rects.length is 4
PASS rects[0].left is 80
FAIL rects[0].top should be 411. Was 407.
PASS rects[0].width is 324
FAIL rects[0].height should be 18. Was 17.
PASS rects[1].left is 8
FAIL rects[1].top should be 451. Was 447.
PASS rects[1].width is 398
FAIL rects[1].height should be 18. Was 17.
PASS rects[2].left is 8
FAIL rects[2].top should be 491. Was 487.
PASS rects[2].width is 360
FAIL rects[2].height should be 18. Was 17.
PASS rects[3].left is 8
FAIL rects[3].top should be 531. Was 527.
PASS rects[3].width is 88
FAIL rects[3].height should be 18. Was 17.
Test 4
PASS rects.length is 7
PASS rects[0].left is 8
FAIL rects[0].top should be 596. Was 590.
PASS rects[0].width is 400
PASS rects[0].height is 278
PASS rects[1].left is 8
FAIL rects[1].top should be 607. Was 601.
PASS rects[1].width is 396
FAIL rects[1].height should be 18. Was 17.
PASS rects[2].left is 8
FAIL rects[2].top should be 725. Was 719.
PASS rects[2].width is 242
FAIL rects[2].height should be 18. Was 17.
PASS rects[3].left is 326
FAIL rects[3].top should be 725. Was 719.
PASS rects[3].width is 44
FAIL rects[3].height should be 18. Was 17.
PASS rects[4].left is 8
FAIL rects[4].top should be 765. Was 759.
PASS rects[4].width is 341
FAIL rects[4].height should be 18. Was 17.
PASS rects[5].left is 8
FAIL rects[5].top should be 805. Was 799.
PASS rects[5].width is 366
FAIL rects[5].height should be 18. Was 17.
PASS rects[6].left is 8
FAIL rects[6].top should be 845. Was 839.
PASS rects[6].width is 67
FAIL rects[6].height should be 18. Was 17.
Test 5
PASS rects.length is 5
PASS rects[0].left is 80
FAIL rects[0].top should be 921. Was 913.
PASS rects[0].width is 324
FAIL rects[0].height should be 18. Was 17.
PASS rects[1].left is 8
FAIL rects[1].top should be 1039. Was 1031.
PASS rects[1].width is 242
FAIL rects[1].height should be 18. Was 17.
PASS rects[2].left is 250
FAIL rects[2].top should be 950. Was 942.
PASS rects[2].width is 76
PASS rects[2].height is 103
PASS rects[3].left is 326
FAIL rects[3].top should be 1039. Was 1031.
PASS rects[3].width is 44
FAIL rects[3].height should be 18. Was 17.
PASS rects[4].left is 8
FAIL rects[4].top should be 1079. Was 1071.
PASS rects[4].width is 284
FAIL rects[4].height should be 18. Was 17.
Test 6
PASS rects.length is 7
PASS rects[0].left is 8
FAIL rects[0].top should be 1224. Was 1214.
PASS rects[0].width is 400
PASS rects[0].height is 200
PASS rects[1].left is 8
FAIL rects[1].top should be 1235. Was 1225.
PASS rects[1].width is 396
FAIL rects[1].height should be 18. Was 17.
PASS rects[2].left is 8
FAIL rects[2].top should be 1275. Was 1265.
PASS rects[2].width is 57
FAIL rects[2].height should be 18. Was 17.
PASS rects[3].left is 105
FAIL rects[3].top should be 1275. Was 1265.
PASS rects[3].width is 44
FAIL rects[3].height should be 18. Was 17.
PASS rects[4].left is 8
FAIL rects[4].top should be 1315. Was 1305.
PASS rects[4].width is 387
FAIL rects[4].height should be 18. Was 17.
PASS rects[5].left is 8
FAIL rects[5].top should be 1355. Was 1345.
PASS rects[5].width is 397
FAIL rects[5].height should be 18. Was 17.
PASS rects[6].left is 8
FAIL rects[6].top should be 1395. Was 1385.
PASS rects[6].width is 171
FAIL rects[6].height should be 18. Was 17.
Test 7
PASS rects.length is 7
PASS rects[0].left is 8
FAIL rects[0].top should be 1471. Was 1459.
PASS rects[0].width is 396
FAIL rects[0].height should be 18. Was 17.
PASS rects[1].left is 8
FAIL rects[1].top should be 1500. Was 1488.
PASS rects[1].width is 400
PASS rects[1].height is 40
PASS rects[2].left is 8
FAIL rects[2].top should be 1511. Was 1499.
PASS rects[2].width is 57
FAIL rects[2].height should be 18. Was 17.
PASS rects[3].left is 105
FAIL rects[3].top should be 1511. Was 1499.
PASS rects[3].width is 44
FAIL rects[3].height should be 18. Was 17.
PASS rects[4].left is 8
FAIL rects[4].top should be 1551. Was 1539.
PASS rects[4].width is 387
FAIL rects[4].height should be 18. Was 17.
PASS rects[5].left is 8
FAIL rects[5].top should be 1591. Was 1579.
PASS rects[5].width is 397
FAIL rects[5].height should be 18. Was 17.
PASS rects[6].left is 8
FAIL rects[6].top should be 1631. Was 1619.
PASS rects[6].width is 171
FAIL rects[6].height should be 18. Was 17.
Test 8
PASS rects.length is 4
FAIL rects[0].left should be 168. Was 169.
FAIL rects[0].top should be 1678. Was 1664.
FAIL rects[0].width should be 269. Was 268.
FAIL rects[0].height should be 316. Was 315.
FAIL rects[1].left should be 137. Was 138.
FAIL rects[1].top should be 1704. Was 1690.
FAIL rects[1].width should be 271. Was 270.
FAIL rects[1].height should be 317. Was 316.
PASS rects[2].left is 107
FAIL rects[2].top should be 1729. Was 1715.
PASS rects[2].width is 246
FAIL rects[2].height should be 289. Was 288.
FAIL rects[3].left should be 76. Was 77.
FAIL rects[3].top should be 1755. Was 1741.
FAIL rects[3].width should be 212. Was 211.
FAIL rects[3].height should be 247. Was 246.
Test 9
FAIL rects.length should be 1. Was 0.
FAIL rects[0].left should be 8. Threw exception TypeError: undefined is not an object (evaluating 'rects[0].left')
FAIL rects[0].top should be 1903. Threw exception TypeError: undefined is not an object (evaluating 'rects[0].top')
FAIL rects[0].width should be 0. Threw exception TypeError: undefined is not an object (evaluating 'rects[0].width')
FAIL rects[0].height should be 18. Threw exception TypeError: undefined is not an object (evaluating 'rects[0].height')
Test 9b
PASS rects.length is 1
PASS rects[0].left is 8
FAIL rects[0].top should be 1903. Was 1887.
PASS rects[0].width is 0
FAIL rects[0].height should be 18. Was 17.
Test 10
PASS rects.length is 4
PASS rects[0].left is 19
FAIL rects[0].top should be 2088. Was 2070.
FAIL rects[0].width should be 18. Was 17.
PASS rects[0].height is 393
PASS rects[1].left is 59
FAIL rects[1].top should be 2088. Was 2070.
FAIL rects[1].width should be 18. Was 17.
PASS rects[1].height is 377
PASS rects[2].left is 99
FAIL rects[2].top should be 2088. Was 2070.
FAIL rects[2].width should be 18. Was 17.
PASS rects[2].height is 372
PASS rects[3].left is 139
FAIL rects[3].top should be 2088. Was 2070.
FAIL rects[3].width should be 18. Was 17.
PASS rects[3].height is 360
Test 11
PASS rects.length is 2
PASS rects[0].left is 8
FAIL rects[0].top should be 2524. Was 2504.
PASS rects[0].width is 400
PASS rects[0].height is 40
PASS rects[1].left is 8
FAIL rects[1].top should be 2535. Was 2515.
PASS rects[1].width is 177
FAIL rects[1].height should be 18. Was 17.
Test 12
PASS rects.length is 1
PASS rects[0].left is 8
FAIL rects[0].top should be 2760. Was 2738.
PASS rects[0].width is 400
PASS rects[0].height is 160
Test 13
PASS rects.length is 4
PASS rects[0].left is 8
FAIL rects[0].top should be 2967. Was 2943.
PASS rects[0].width is 396
FAIL rects[0].height should be 18. Was 17.
PASS rects[1].left is 8
FAIL rects[1].top should be 3007. Was 2983.
PASS rects[1].width is 398
FAIL rects[1].height should be 18. Was 17.
PASS rects[2].left is 8
FAIL rects[2].top should be 3047. Was 3023.
PASS rects[2].width is 360
FAIL rects[2].height should be 18. Was 17.
PASS rects[3].left is 8
FAIL rects[3].top should be 3087. Was 3063.
PASS rects[3].width is 306
FAIL rects[3].height should be 18. Was 17.
Test 14
PASS rects.length is 0
PASS successfullyParsed is true

TEST COMPLETE
